rename: #35 Andy 폴더를 분리, 각 GroupVote 디렉토리로 배치
