+++ moyeo/moyeo/Presentation/View/MainView.swift
@@ -20,7 +20,6 @@ struct MainView: View {
     // Andy
     @StateObject var sharedDm = SharedDateModel()
     
-    
     var body: some View {
         VStack {
             
@@ -68,7 +67,7 @@ struct MainView: View {
             .background(Color.black)
             .cornerRadius(10)
             .sheet(isPresented: $isPresentingGroupSetView) {
-                AndyGroupSetView(isPresentingGroupSetView: $isPresentingGroupSetView, sharedDm: sharedDm)
+                GroupSetView(isPresentingGroupSetView: $isPresentingGroupSetView, sharedDm: sharedDm)
             }
             
         }
@@ -76,7 +75,7 @@ struct MainView: View {
         .navigationDestination(for: MainRoute.self) { destination in
             switch destination {
             case .groupVoteView:
-                AndyGroupVoteView(sharedDm: sharedDm)
+                GroupVoteView(sharedDm: sharedDm)
             case .groupResultView:
                 GroupResultView()
             }
