+++ moyeo/moyeo/Shared/GroupVote/AlertError.swift
