+++ moyeo/moyeo/Shared/GroupVote/EventStoreError.swift
