+++ moyeo/moyeo/Shared/GroupVote/EventStoreManager+FullAccess.swift
